refactor platform interface
