+++ internal/platform/config.go
@@ -1,0 +1,5 @@
+package platform
+
+type Config struct {
+	UserAgent string
+}
